refactor: expose la définition d'une dépense individuelle
